Based on MiFi Controller fork.
Added ARC;
Use new gamepad UI on iPhone.

+++ dospad/Shared/DosPadUIApplication.h
@@ -10,6 +10,6 @@
 
 @interface DosPadUIApplication : UIApplication
 {
-	NSInteger lastEventFlags;
+	int lastEventFlags;
 }
 @end
